uncomment permissions

+++ src/com/seitptt/model/Core.java
@@ -155,32 +155,32 @@ public class Core {
 	}
 
 	public void approveTeachingRequest(TeachingRequest tr) {
-		//this.checkPermission(PTTDirector.class);
+		this.checkPermission(PTTDirector.class);
 		tr.approve();
 	}
 
 	public void removeTeachingRequirement(TeachingRequirement tr) {
-		//this.checkPermission(ClassDirector.class);
+		this.checkPermission(ClassDirector.class);
 		Database.removeTeachingRequirementFromDB(tr);
 	}
 
 	public void removeTeachingRequest(TeachingRequest tr) {
-		//this.checkPermission(ClassDirector.class);
+		this.checkPermission(ClassDirector.class);
 		Database.removeTeachingRequestFromDB(tr);
 	}
 
 	public void organiseTraining(Teacher t) {
-		//this.checkPermission(Administrator.class);
+		this.checkPermission(Administrator.class);
 		t.train();
 	}
 
 	public void createAndAddTeachingRequest(Teacher t, Classes c, TeachingRequirement tr) {
-		//this.checkPermission(Administrator.class);
+		this.checkPermission(Administrator.class);
 		new TeachingRequest(t, c, tr);
 	}
 
 	public ListOfTeachingRequirements getListOfTeachingRequirements() {
-		//this.checkPermission(ClassDirector.class);
+		this.checkPermission(ClassDirector.class);
 		return Database.getTeachingRequirementsFromDB();
 	}
 
@@ -189,7 +189,7 @@ public class Core {
 	}
 
 	public ListOfEmployees getListOfTeachers() {
-		//this.checkPermission(Administrator.class);
+		this.checkPermission(Administrator.class);
 		return Database.getEmployeesFromDB().getTeachers();
 	}
